+++ changes.md
@@ -7,6 +7,7 @@
 4. Support for TH2 'box' with negative values (crosses) and 'box1' (buttons) 
 5. Support 'same' for 'box' and 'col' draw options 
 6. Implement drawing of TProfile2D
+7. Support logx/y/z, gridx/y, tickx/y draw options for TPad/TCanvas
 
 
 ## Changes in 4.7.0